Extract sign up into it's own spec file
Ahead of addition variations on this spec.
