+++ spec/features/manage_alerts_spec.rb
@@ -1,32 +1,6 @@
 require 'spec_helper'
 
 feature "Manage alerts" do
-  # In order to see new development applications in my suburb
-  # I want to sign up for an email alert
-  
-  scenario "Sign up for email alert" do
-    VCR.use_cassette('planningalerts') do
-      visit '/alerts/signup'
-
-      fill_in("alert_email", with: "example@example.com")
-      fill_in("alert_address", with: "24 Bruce Road, Glenbrook")
-      click_button("Create alert")
-    end
-
-    page.should have_content("Now check your email")
-
-    unread_emails_for("example@example.com").size.should == 1
-    open_email("example@example.com")
-    current_email.should have_subject("Please confirm your planning alert")
-    current_email.default_part_body.to_s.should include("24 Bruce Road, Glenbrook NSW 2773")
-    click_first_link_in_email
-
-    page.should have_content("your alert has been activated")
-    page.should have_content("24 Bruce Road, Glenbrook NSW 2773")
-    page.should_not have_content("You now have several email alerts")
-    Alert.active.find_by(address: "24 Bruce Road, Glenbrook NSW 2773", radius_meters: "2000", email: current_email_address).should_not be_nil
-  end
-
   scenario "Unsubscribe from an email alert" do
     # Adding arbitrary coordinates so that geocoding is not carried out
     alert = create(:alert, address: "24 Bruce Rd, Glenbrook", email: "example@example.com",
